Struct chunked::ChunkedVec

pub struct ChunkedVec<Item: Clone> { /* fields omitted */ }

A vector which stores items of a known type in a SizedChunkedArena

Methods

impl<Item: Clone> ChunkedVec<Item>

fn new(chunker: Box<Chunker>) -> Self

Create a new chunked vector

fn len(&self) -> usize

Get the number of elements in the vector

fn is_empty(&self) -> bool

Is the chunked vector empty?

fn at(&self, index: usize) -> &Item

Get a reference to the item at index

fn at_mut(&mut self, index: usize) -> &mut Item

Get a mutable reference to the item at index

fn push(&mut self, item: Item)

Push an item onto the vector

fn pop(&mut self) -> Option<Item>

Remove and return the last item, if the vector wasn't empty