Struct chunked::MultiSized

pub struct MultiSized<Bin: SizedChunkedCollection> {
    pub bins: Vec<Bin>,
    // some fields omitted
}

Based on a collection type for fixed-size items ("Bin"), creates a collection for heterogenously-sized items which will be stored in the most appropriately-sized bin.

All Bins will use children of a main chunker to create their chunks.

Fields
bins: Vec<Bin>

All fixed-size bins in this multi-sized collection

The bin at index i will have item-size base_size * 2 ^ i

Methods

impl<B: SizedChunkedCollection> MultiSized<B>

fn new(chunker: Box<Chunker>, base_size: usize) -> Self

Create a new MultiSized collection using the given chunker as a main chunker and a base size that represents the smallest expected item size (this will be used as the item size of the smallest-sized Bin)

fn size_to_index(&self, size: usize) -> usize

Get the index of the Bin which stores items of size size

fn bin_for_size_mut(&mut self, size: usize) -> &mut B

Get a reference to the Bin most appropriately sized given item size size