Trait chunked::Chunker

pub trait Chunker {
    fn chunk_size(&self) -> usize;
    fn name(&self) -> &str;
    fn with_chunk_size(&self, size: usize) -> Box<Chunker>;
    fn with_name(&self, name: &str) -> Box<Chunker>;
    fn child(&self, suffix: &str) -> Box<Chunker>;
    fn create_chunk(&mut self) -> *mut u8;
    unsafe fn destroy_chunk(&mut self, ptr: *mut u8);

    fn load_chunk(&mut self, _index: usize) -> *mut u8 { ... }
}

Describes a strategy for creation and destruction of chunks, adhering to configurable settings

Required Methods

fn chunk_size(&self) -> usize

Get the set chunk size

fn name(&self) -> &str

Get the set chunk name root

This name can be used internally to store a chunk in a specific file.

fn with_chunk_size(&self, size: usize) -> Box<Chunker>

Create an identical chunker, but with different chunk size

fn with_name(&self, name: &str) -> Box<Chunker>

Create an identical chunker, but with a different name

fn child(&self, suffix: &str) -> Box<Chunker>

Create an identical chunker, with the name of the current chunker extended by a suffix.

This can be used to store subproperties of complex objects in similar-sounding files

fn create_chunk(&mut self) -> *mut u8

Create a new chunk with the set chunk size

unsafe fn destroy_chunk(&mut self, ptr: *mut u8)

Destroys a chunk that was created by this chunker as well as any persisted representation of the chunk.

Undefined behaviour if passed a string that was not created by create_chunk or load_chunk

Provided Methods

fn load_chunk(&mut self, _index: usize) -> *mut u8

Load a persisted chunk that was previously created by this chunker, given an index to identify the particular chunk. Note: the default implementation just creates a new, empty chunk